Refactor goal creation in contexts
The reason for this is that some custom goals, e.g., the open rave knee goal, return a special
sample that is known to be valid the first time a goal is sampled. Every time a new problem
definition is created, a new goal has to be created as well, as there is no other way to tell the
goal when it is used in a new context.