Use std::high_resolution_clock if it is steady
There are two fundamentally different types of clocks, called steady and
unsteady. Steady clocks are monotonically increasing, which makes them good at
measuring duration. Unsteady clocks sometimes move backward in time to correct
for accumulated drift, which makes them good for telling the exact time. We
need a steady clock because we care about measuring duration.

The downside of std::steady_clock is that it sometimes has a lower resolution
than std::high_resolution_clock, so we should only use it if the high
resolution version is not steady.