Specify third-party includes as system includes
Our tools are compiled with stringent warnings and every warning is treated as
an error. Some of the libraries we depend on do not compile without warnings,
and since warnings are treated as errors it prevents our tools from
buildings.

The best way to fix this would be to fix the libraries we depend on.

The second best way I know of is to specify the include directories of these
libraries as system includes, which tells the compiler not to report warnings
for these headers.